btrfs-progs: send-stream: use proper type for cmd in read_cmd
We just read and assign the value, there's no reason to use different
types.

+++ send-stream.c
@@ -72,7 +72,7 @@ out:
 static int read_cmd(struct btrfs_send_stream *sctx)
 {
 	int ret;
-	int cmd;
+	u16 cmd;
 	u32 cmd_len;
 	char *data;
 	u32 pos;
